refactored with composition
