+++ src/components/uploader/main.jsx
@@ -1,9 +1,15 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import axios from "axios";
 
-export default function Uploader() {
+export default function Uploader(props) {
+  const [mode, setMode] = useState("");
+  useEffect(() => {
+    setMode(props.mode);
+    console.log("Oooo", mode);
+  });
+
   const [selectedFile, setSelectedFile] = useState("");
-  const [isSelected, setSelected] = useState(false);
+  const [isSelected, setSelected] = useState(false); // TODO: if selected make clickable upload button
   const [imgLink, setImgLink] = useState(null);
 
   const changeHandler = (e) => {
@@ -52,7 +58,7 @@ export default function Uploader() {
             display: flex;
           }
           .img {
-            display: ${imgLink ? true : 'none'};
+            display: ${imgLink ? true : "none"};
             max-width: 400px;
             max-heigh: 400px;
             border-style: solid;
@@ -66,18 +72,31 @@ export default function Uploader() {
       </style>
       <div className="flexy">
         <div>
-          <label className="custom_file_btn">
-            <div>selected file</div>
-            <input className="hide" type="file" onChange={changeHandler} />
-          </label>
+          {console.log("mode", mode)}
+          {mode === "img" || mode == "gltf" ? (
+            <div>
+              <label className="custom_file_btn">
+                <div>select file</div>
+                <input className="hide" type="file" onChange={changeHandler} />
+              </label>
+            </div>
+          ) : (
+            <div></div>
+          )}
+
           <button className="custom_btn" onClick={handleSubmission}>
             upload
           </button>
         </div>
 
-        <div className="vertical_alignment">
-          <div className="result">{selectedFile.name}</div>
-        </div>
+        {mode === "img" || mode == "gltf" ? (
+          <div className="vertical_alignment">
+            <div className="result">{selectedFile.name}</div>
+          </div>
+        ) : (
+          <div />
+        )}
+        
       </div>
       <div className="div-pad-top">
         <img className="img" src={imgLink ? imgLink : null}></img>
